Addition of feature to create the next file
Creates the next file that can be supported, in other words chooses a name that hasn´t been used yet "Test_x" or "Fligh_x". The x can be replaced by the corresponding number.

Also, the program by a input pin is able to determine if the user wants a new test or flight file. It handles directories to maintain everything in their own folders.